requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Task queue
celery[redis]==5.3.4
redis==4.6.0
flower==2.0.1
kombu==5.3.4

# AI/ML services
openai==1.3.5
google-cloud-texttospeech==2.14.2
elevenlabs==0.2.26

# Media processing
ffmpeg-python==0.2.0
Pillow==10.1.0
moviepy==1.0.3
numpy==1.24.3

# AWS
boto3==1.29.7
botocore==1.32.7

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.25.2
tenacity==8.2.3
structlog==23.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring & Logging
prometheus-client==0.19.0
